backend/requirements.txt
flask==3.0.0
Werkzeug==3.0.1
authlib==1.3.0
requests==2.31.0
python-dotenv==1.0.0
